+++ src/StatPeak/StatPeakServerUtil.cs
@@ -9,6 +9,7 @@ namespace StatPeak;
 
 internal class StatPeakServerUtil
 {
+    // Acts as a filter and a translator. All stats listed here, will be sent to remote server
     internal static Dictionary<string, string> ExternalKeyMap = new Dictionary<string, string>()
     {
         // { IncomingKey, OutgoingKey }
@@ -18,6 +19,8 @@ internal class StatPeakServerUtil
         { Stat.DistanceWhileAirborne, "fallen" },
         { Stat.DistanceClimbedOnVines, "slid" },
         { Stat.TotalJumps, "jumps" },
+        { Stat.GreatestContinuousClimb, "greatest_climb" },
+
         { "Hunger", "hunger" },
         { "Poison", "poison" },
         { "Cold", "cold" },
@@ -25,6 +28,7 @@ internal class StatPeakServerUtil
         { "Drowsy", "sleep" },
         { "Curse", "curse" },
         { "Injury", "injury" },
+
         { Stat.TotalFaints, "faints" },
         { Stat.TotalDeaths, "deaths" },
         { Stat.TotalRevives, "revives" },
@@ -32,6 +36,77 @@ internal class StatPeakServerUtil
         { Stat.ItemsThrown, "items_thrown" },
         { Stat.ItemsGrabbed, "items_grabbed" },
         { Stat.ItemsCooked, "items_cooked" },
+
+        { ItemName.YellowCrispberry, "eaten_y_crispberry" },
+        { ItemName.RedCrispberry, "eaten_r_crispberry" },
+        { ItemName.GreenCrispberry, "eaten_g_crispberry" },
+        { ItemName.YellowClusterberry, "eaten_y_clusterberry" },
+        { ItemName.BlackClusterberry, "eaten_b_clusterberry" },
+        { ItemName.RedClusterberry, "eaten_r_clusterberry" },
+        { ItemName.GreenClusterberry, "eaten_g_clusterberry" },
+        { ItemName.YellowKingberry, "eaten_y_kingberry" },
+        { ItemName.GreenKingberry, "eaten_g_kingberry" },
+        { ItemName.PurpleKingberry, "eaten_p_kingberry" },
+        { ItemName.BrownBerrynana, "eaten_br_berrynana" },
+        { ItemName.BlueBerrynana, "eaten_bl_berrynana" },
+        { ItemName.PinkBerrynana, "eaten_p_berrynana" },
+        { ItemName.YellowBerrynana, "eaten_y_berrynana" },
+        { ItemName.YellowWinterberry, "eaten_y_winterberry" },
+        { ItemName.OrangeWinterberry, "eaten_o_winterberry" },
+        { ItemName.Napberry, "eaten_napberry" },
+        { ItemName.MedicinalRoot, "eaten_medroot" },
+        { ItemName.Egg, "eaten_egg" },
+        { ItemName.Bird, "eaten_chicken" },
+        { ItemName.Marshmallow, "eaten_marshmallow" },
+        { ItemName.CoconutHalf, "eaten_coconut" },
+        { ItemName.Honeycomb, "eaten_honeycomb" },
+        { ItemName.ButtonShroom, "eaten_buttonshroom" },
+        { ItemName.BugleShroom, "eaten_bugleshroom" },
+        { ItemName.ClusterShroom, "eaten_clustershroom" },
+        { ItemName.ChubbyShroom, "eaten_chubbyshroom" },
+        { ItemName.Tick, "eaten_tick" },
+        { ItemName.TrailMix, "eaten_trailmix" },
+        { ItemName.GranolaBar, "eaten_granola" },
+        { ItemName.SportsDrink, "eaten_sportsdrink" },
+        { ItemName.EnergyDrink, "eaten_energy" },
+        { ItemName.AirlineFood, "eaten_airlinefood" },
+        { ItemName.ScoutCookies, "eaten_cookies" },
+        { ItemName.BigLollipop, "eaten_lollipop" },
+
+        // TODO: Check tracking for all these
+        { ItemName.PandorasLunchbox, "used_lunchbox" },
+        { ItemName.CureAll, "used_cureall" },
+        { ItemName.CursedSkull, "used_cursedskull" },
+        { ItemName.ScoutEffigy, "used_scouteffigy" },
+        { ItemName.Balloon, "used_balloon" },
+        { ItemName.Sunscreen, "used_sunscreen" },
+        { ItemName.ScoutCannon, "used_cannon" },
+        { ItemName.Blowgun, "used_blowgun" },
+        { ItemName.Piton, "used_piton" },
+        { ItemName.FirstAidKit, "used_medkit" },
+        { ItemName.HeatPack, "used_heatpack" },
+        { ItemName.Antidote, "used_antidote" },
+        { ItemName.Bandages, "used_bandages" },
+        { ItemName.FriendshipBugle, "used_friendshipbugle" },
+        { ItemName.Bugle, "used_bugle" },
+        { ItemName.PortableStove, "used_stove" },
+        { ItemName.Binoculars, "used_binoculars" },
+        { ItemName.Flare, "used_flares" },
+        { ItemName.RopeSpool, "used_ropespool" },
+        { ItemName.AntiropeSpool, "used_antiropespool" },
+        { ItemName.RopeCannon, "used_ropegun" },
+        { ItemName.AntiropeCannon, "used_antiropegun" },
+        { ItemName.ChainLauncher, "used_chaingun" },
+
+        { Stat.RopePlaced, "placed_rope" },
+        { Stat.ChainPlaced, "placed_chain" },
+        { Stat.AntiropePlaced, "placed_antirope" },
+
+        { Stat.HasEscaped, "escaped" },
+        { Stat.HasFailed, "failed" },
+
+        { Stat.CampfiresLit, "campfires_lit" },
+        { Stat.TicksRemoved, "ticks_removed" },
     };
 
     public static Dictionary<string, double> RemapKeys(Dictionary<string, double> map)